BootService Starter is modified Date time is added new functin for millisecond to date time conversion Listerner service is updated for messages comming from watch Send message service is udated to update DataAPI Room Selector is updated to refresh DataApi on room change, refresh measurements Strings.xml is modified by adding new data paths
Wear:
Date time is added new function to returen day of the week
gridview_pageritem.xml is added for the main activity which provides a custom view for grid pager
rect_activities are prepared for recangular watches

MeasurementDetail activity is updated to with two charts, barchart and linechart, bar chart shows the average values of a selected measurement through the last 6 days and linechart shows the trend of consumption changes of current day.
Strings. xml is updated by Tag_ids to support setTag function of  custom views