Alarm manager service runs the Async task to check the tasks by interval.
And class updated "And" to "AND" in rules.

Greater class added to check the temperature, humididty and co2 measurements by numbers
LessThan class added to check the temperature, himidity, and co2 measurements by their numerical values

Json_Ruler is a updated and added by alarm manager to start service priodically to check the rules
Json_Ruler fields are updated, json rules json updated also, rule interval separated to check the rules as a hole in each interval

manifest file is updated to support Alarm manager and also to register the service that will be called by alarm manager

Strings.xml updated by measurements that can be used to evaluate the rules and they are used to generate rule bindings

RuleFactGenerator updated to generate rule bindings.